requirements.txt
absl-py==0.6.1
astor==0.7.1
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
gast==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-speech==0.36.2
google-cloud-texttospeech==0.3.0
google-cloud-translate==1.3.3
googleapis-common-protos==1.5.6
grpcio==1.17.1
h5py==2.9.0
idna==2.8
imageio==2.4.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
Pillow==5.4.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
sounddevice==0.3.12
SoundFile==0.10.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
